requirements.txt
discord.py
python-dotenv
aiosqlite
ezcord
